@@ -3,33 +3,60 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@ably-labs/react-hooks": "^2.1.0",
+    "@apollo/client": "^3.7.11",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@giphy/js-fetch-api": "^4.1.2",
     "@giphy/react-components": "^5.4.0",
     "@n8tb1t/use-scroll-position": "^2.0.3",
-    "@prismicio/client": "^6.0.0",
-    "@prismicio/react": "^2.0.3",
-    "@react-hook/mouse-position": "^4.1.3",
+    "@prismicio/client": "^6.7.3",
+    "@prismicio/react": "^2.5.2",
+    "@prismicio/richtext": "^2.1.5",
+    "@reach/dialog": "^0.18.0",
+    "@reach/tabs": "^0.18.0",
+    "@reach/visually-hidden": "^0.18.0",
+    "@sanity/cli": "^3.10.0",
+    "@sanity/client": "^6.0.1",
+    "@szhsin/react-accordion": "^1.2.0",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
     "@widgetbot/react-embed": "^1.4.0",
+    "ably": "^1.2.29",
+    "boring-avatars": "^1.10.1",
+    "classnames": "^2.3.2",
     "clsx": "^1.1.1",
     "dayjs": "^1.10.7",
-    "framer-motion": "^6.0.0",
+    "dotenv": "^16.0.1",
+    "framer-motion": "^10.12.16",
+    "graphql": "^16.6.0",
+    "graphql.macro": "^1.4.2",
     "hamburger-react": "^2.4.1",
     "immer": "^9.0.12",
-    "react": "^17.0.2",
+    "js-cookie": "^3.0.1",
+    "lodash": "^4.17.21",
+    "nanoid": "^4.0.2",
+    "netlify-identity-widget": "^1.9.2",
+    "react": "^18.2.0",
+    "react-debounce-input": "^3.3.0",
     "react-device-detect": "^2.1.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-double-marquee": "^1.1.0",
-    "react-icons": "^4.3.1",
+    "react-icons": "^4.10.1",
     "react-intersection-observer": "^8.33.1",
+    "react-netlify-identity": "^0.2.7",
+    "react-responsive": "^9.0.2",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
+    "react-select": "^5.7.0",
+    "react-slider": "^2.0.4",
     "react-spinners": "^0.11.0",
+    "react-tabs": "^6.0.2",
     "react-use-mounted": "^2.0.0",
+    "swiper": "^9.2.0",
+    "swr": "^2.1.2",
+    "usehooks-ts": "^2.9.1",
     "web-vitals": "^2.1.0",
     "zustand": "^3.6.9"
   },
@@ -56,5 +83,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "simple-zustand-devtools": "^1.1.0"
   }
 }
