@@ -1,134 +1,163 @@
 import styled from "@emotion/styled";
-import React, { useRef, useState } from "react";
-import { isDesktop } from "react-device-detect";
-import {
-  BsPause,
-  BsPlay,
-  BsVolumeDown,
-  BsVolumeMute,
-  BsVolumeUp
-} from "react-icons/bs";
-import config from "../../config";
+import React, { useEffect, useRef } from "react";
+import { GoPlay, GoSquareFill } from "react-icons/go";
 import useAudioPlayerStore from "../../Stores/AudioPlayerStore";
-import Loader from "../Loader";
+import useBroadcastStore from "../../Stores/BroadcastStore";
+import { BREAKPOINT_MD, BREAKPOINT_XS, STREAM_URL } from "../../config";
+import Arrow from "../../images/Arrow";
+import Dot from "../../images/Dot";
+import DotGrey from "../../images/DotGrey";
+import StreamShortInfo from "./StreamShortInfo";
 
 const Container = styled.div`
-  display: flex;
-  margin-left: 1rem;
-  align-items: center;
-`;
-const ControlButton = styled.button`
-  color: var(--color);
-  border: 0;
-  border-radius: 0.25rem;
-  height: 3rem;
-  padding: 0 2rem;
+padding: 0 2rem;
+@media (max-width: ${BREAKPOINT_XS}px) {
+  padding: 0 1rem;
+}
+
+button {
   cursor: pointer;
-  font-weight: bold;
-  line-height: 0.75rem;
-  &:hover {
-    color: var(--second);
+}
+> header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  overflow: hidden;
+  gap: 0rem;
+  line-height: 3rem;
+  @media (max-width: ${BREAKPOINT_MD}px) {
+    display: flex;
+  } 
+}
+
+.status {
+  color: var(--second);
+  box-sizing: border-box;
+  text-transform: uppercase;
+  white-space: nowrap;
+  display: flex;
+  height: 100%;
+  justify-content: center;
+  align-items: center;
+  svg {
+    height: 1.25rem;
+    width: 1.25rem;
   }
-`;
-const PlayButton = styled(ControlButton)`
-  position: relative;
-  padding: 0;
-  padding: 0;
-  background: transparent;
-  font-size: 1.5rem;
-`;
-const VolumeButton = styled(PlayButton)``;
-
-const VolumeSlider = styled.div`
-  position: absolute;
-  top: 3rem;
-  input {
-    writing-mode: bt-lr; /* IE */
-    -webkit-appearance: slider-vertical; /* WebKit */
-    width: 1rem;
-    height: 6rem;
-    padding: 0 0.25rem;
-  }
-`;
-
-const AudioPlayer = () => {
-  const [volume, setVolume] = useState(0.5);
-  const [showVolumeSlider, setShowVolumeSlider] = useState(false);
-  const volumeSlider = useRef();
-  const audioPlayer = useRef();
-  // const source = useRef();
-  // const [track, setTrack] = useState(config.STREAM_URL);
-  const { isPlaying, setIsPlaying, isLoading, setIsLoading } =
-    useAudioPlayerStore();
-
-  const togglePlay = () => {
-    if (isPlaying) {
-      audioPlayer.current.pause();
-      setIsPlaying(false)
-    } else {
-      // setTrack(config.STREAM_URL);
-      setIsPlaying(true)
-      audioPlayer.current.play();
+  @media (max-width: ${BREAKPOINT_MD}px) {
+    .appendix {
+      display: none;
     }
-  };
-  const toggleVolumeSlider = () => {
-    setShowVolumeSlider(!showVolumeSlider);
-  };
-  const changeVolume = (e) => {
-    setVolume(e.target.value);
-    audioPlayer.current.volume = e.target.value;
-  };
+  }
+  @media (max-width: ${BREAKPOINT_MD}px) {
+    .appendix {
+      display: none;
+    }
+  }
+  @media (max-width: ${BREAKPOINT_XS}px) {
+    .now {
+      display: none;
+    }
+  }
+}
+`;
+export const PlayButton = styled.button`
+  height: 3rem;
+  display: flex;
+  align-items: center;
+  justify-content: end;
+  padding: 0;
+  @media (max-width: ${BREAKPOINT_MD}px) {
+    justify-content: flex-start;
+  }
+  @media (max-width: ${BREAKPOINT_XS}px) {
+    width: 2rem;
+  }
+  svg {
+    width: 2rem;
+    height: 1.5rem;
+  }
+`;
 
-  const onPlaying = () => {
-    // TODO: visualisation
-  };
+const AudioPlayer = ({ isExpanded, setIsExpanded }) => {
 
-  const onCanPlay = () => {
+  const { isPlaying, setIsPlaying, setIsLoading, isLoading, volume } =
+    useAudioPlayerStore();
+  const { canPlay, setCanPlay, setIsPlaying: setBroadcastIsPlaying, isStreaming, isLive } = useBroadcastStore()
+
+  let audioPlayer = useRef();
+
+  const play = () => {
+    setIsPlaying(true)
+    setBroadcastIsPlaying(false)
+    audioPlayer.current.play();
+  }
+
+  const pause = () => {
+    setIsPlaying(false);
+    audioPlayer.current.pause();
+  }
+
+  const handleEnded = () => {
+    console.warn('Audio Stream ended.')
+    setCanPlay(false);
+  }
+
+  const handleError = (e) => {
+    console.warn('No Stream available', e.target.error)
+    setIsLoading(false)
+    setCanPlay(false);
+  }
+
+  const onCanPlayThrough = () => {
+    setCanPlay(true);
     setIsLoading(false);
     audioPlayer.current.volume = volume;
   };
 
+  useEffect(() => {
+    audioPlayer.current.volume = volume;
+  }, [volume])
+
+  useEffect(() => {
+    if (!isPlaying) {
+      pause()
+    }
+  }
+    , [isPlaying])
+
+  const handleClick = () => { setIsExpanded(!isExpanded) }
+
   return (
     <Container>
-      <audio
-        ref={audioPlayer}
-        volume={volume}
-        onTimeUpdate={onPlaying}
-        onCanPlay={onCanPlay}
-        src={config.STREAM_URL}
-      />
-      {isDesktop && (
-        <>
-          <VolumeButton onClick={toggleVolumeSlider}>
-            {volume <= 0 && <BsVolumeMute />}
-            {volume > 0 && volume < 0.6 && <BsVolumeDown />}
-            {volume >= 0.6 && <BsVolumeUp />}
-          </VolumeButton>
-
-          {showVolumeSlider && (
-            <VolumeSlider>
-              <input
-                orient="vertical"
-                type="range"
-                max="1"
-                min="0"
-                value={volume}
-                step="0.01"
-                ref={volumeSlider}
-                onChange={changeVolume}
-              />
-            </VolumeSlider>
+      <header>
+        <audio
+          ref={audioPlayer}
+          volume={volume}
+          onCanPlayThrough={onCanPlayThrough}
+          onEnded={handleEnded}
+          onError={handleError}
+          src={STREAM_URL}
+        />
+        {canPlay ? (<>
+          <div className="status">
+            {isLive() && (<Dot />)}
+            {isStreaming() && (<DotGrey />)}
+          </div>
+          {isPlaying ? (
+            <PlayButton onClick={() => pause()}>
+              <GoSquareFill />
+            </PlayButton>
+          ) : (
+            <PlayButton onClick={() => play()}>
+              <GoPlay />
+            </PlayButton>
           )}
-        </>
-      )}
-
-      <PlayButton onClick={togglePlay}>
-        {isLoading ? (
-          <Loader size={15} />
-        ) : (
-          <>{isPlaying ? <BsPause /> : <BsPlay />}</>
-        )}
-      </PlayButton>
+          <StreamShortInfo />
+          <button onClick={handleClick}>
+            <Arrow flipped={!isExpanded} />
+          </button>
+        </>) : (<>Offline</>)}
+      </header>
     </Container>
   );
 };
