@@ -1,19 +1,16 @@
 import styled from "@emotion/styled";
 import FadeIn from "../Animations/FadeIn";
-import RandomShape from "./RandomShape";
 
-const Container = styled.p`
-  text-align: center;
-  img {
-    margin: 3rem auto;
-    width: 7rem;
-  }
+const Container = styled.div`
+  width: 100%;
+  border-bottom: 2px solid var(--color);
+  margin-bottom: 2rem;
+  margin-top: 2rem;
 `;
 const Divider = () => {
   return (
     <FadeIn>
       <Container>
-        <RandomShape/>
       </Container>
     </FadeIn>
   );
