@@ -1,17 +1,14 @@
 import React from "react";
-import FadeIn from "../Animations/FadeIn";
 
 const KeyFieldParagraph = ({ text }) => {
   if (!text) return <></>;
   return (
     <div>
-      {text?.split("\n").map((line,index) => (
-        <FadeIn key={index}>
-          <span key={index}>
+      {text?.split("\n").map((line, index) => (
+        <span key={index}>
           {line}
           <br />
         </span>
-          </FadeIn>
       ))}
     </div>
   );
