@@ -1,27 +1,18 @@
 import produce from "immer";
+import { mountStoreDevtool } from 'simple-zustand-devtools';
 import create from "zustand";
-import { persist } from "zustand/middleware";
 
 const useThemeStore = create(
-  persist(
-    produce((set) => ({
-      keyword: "vienna",
-      setKeyword: (keyword) => set((state) => (state.keyword = keyword)),
-      mousePosition: {x:0,y:0}, 
-      nightMode: false,
-      setNightMode: (value)=> set((state) => {
-        state.nightMode = value;
-      }),
-      showGifs: true,
-      setShowGifs: (value)=> set((state) => {
-        state.showGifs = value;
-      }),
-      setMousePosition: (pos) => {
-        return set((state) => {
-          state.mousePosition = pos;
-        })},
-    }))
-  )
+  produce((set) => ({
+    nightMode: false,
+    setNightMode: (value) => set({ nightMode: value }),
+    searchbarIsVisible: false,
+    setSearchbarIsVisible: (value) => set({ searchbarIsVisible: value })
+  }))
 );
 
 export default useThemeStore;
+
+if (process.env.NODE_ENV === 'development') {
+  mountStoreDevtool('ThemeStore', useThemeStore);
+}
