@@ -11,142 +11,142 @@ const Grid = () => {
     >
       <path
         d="M481 653.83L187.57 713.02L133.51 723.92L110.37 728.59L56.31 739.5L1.02002 750.65"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1398.06 413.12L1878.22 316.58"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1398.31 654.48L1739.46 723.16"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1.02002 316.32L56.31 327.45L110.37 338.34L133.51 343L187.57 353.88L480.88 412.93"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1368.1 270.61H510.87C494.301 270.61 480.87 284.041 480.87 300.61V755.95C480.87 772.519 494.301 785.95 510.87 785.95H1368.1C1384.67 785.95 1398.1 772.519 1398.1 755.95V300.61C1398.1 284.041 1384.67 270.61 1368.1 270.61Z"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1526.58 785.72V828.13C1526.58 836.086 1523.42 843.717 1517.79 849.343C1512.17 854.969 1504.54 858.13 1496.58 858.13H382.4C374.444 858.13 366.813 854.969 361.187 849.343C355.561 843.717 352.4 836.086 352.4 828.13V315.65"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M352.4 270.78V228.43C352.4 220.473 355.561 212.843 361.187 207.217C366.813 201.591 374.444 198.43 382.4 198.43H1496.58C1504.54 198.43 1512.17 201.591 1517.79 207.217C1523.42 212.843 1526.58 220.473 1526.58 228.43V741.74"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1691.41 315.65V920.74C1691.41 928.696 1688.25 936.327 1682.62 941.953C1677 947.579 1669.37 950.74 1661.41 950.74H217.57C209.614 950.74 201.983 947.579 196.357 941.953C190.731 936.327 187.57 928.696 187.57 920.74V858.13"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M187.57 816.42V135.82C187.57 127.864 190.731 120.233 196.357 114.607C201.983 108.981 209.614 105.82 217.57 105.82H1661.41C1669.37 105.82 1677 108.981 1682.62 114.607C1688.25 120.233 1691.41 127.864 1691.41 135.82V272.3"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1847.98 1H31C14.4315 1 1 14.4315 1 31V1025.56C1 1042.13 14.4315 1055.56 31 1055.56H1847.98C1864.55 1055.56 1877.98 1042.13 1877.98 1025.56V31C1877.98 14.4315 1864.55 1 1847.98 1Z"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M786.36 785.9L626.66 1056.59"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M633.24 785.9L315.04 1055.56"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M492.58 780.06L12.96 1049.48"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1679.44 111.58L1386.74 276.72"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M938.52 785.12L939.49 1055.18"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1092.73 785.72L1252.43 1056.41"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1245.85 785.72L1564.05 1055.38"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M1386.16 780.16L1866.34 1049.35"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M199.16 112.25L492.15 276.91"
-        stroke="var(--color)"
+        stroke="currentColor"
         strokeWidth="1.5"
         strokemitterlimit="10"
       />
       <path
         d="M187.927 859.05L187.57 858.685L187.212 859.051L183.18 863.182V862.477L187.57 857.754L191.96 862.477V863.174L187.927 859.05Z"
-        fill="var(--color)"
-        stroke="var(--color)"
+        fill="currentColor"
+        stroke="currentColor"
       />
       <path
         d="M352.757 315.27L352.4 314.905L352.042 315.271L348.01 319.402V318.697L352.4 313.974L356.79 318.697V319.394L352.757 315.27Z"
-        fill="var(--color)"
-        stroke="var(--color)"
+        fill="currentColor"
+        stroke="currentColor"
       />
       <path
         d="M1691.05 272.92L1691.41 273.286L1691.77 272.919L1695.8 268.788V269.494L1691.41 274.216L1687.02 269.494V268.796L1691.05 272.92Z"
-        fill="var(--color)"
-        stroke="var(--color)"
+        fill="currentColor"
+        stroke="currentColor"
       />
       <path
         d="M1526.22 740.88L1526.58 741.245L1526.94 740.879L1530.97 736.748V737.454L1526.58 742.176L1522.19 737.454V736.765L1526.22 740.88Z"
-        fill="var(--color)"
-        stroke="var(--color)"
+        fill="currentColor"
+        stroke="currentColor"
       />
       <path
         d="M1739.51 723.379L1739.92 723.071L1739.6 722.67L1736.02 718.142L1736.72 718.231L1740.85 723.186L1735.6 726.939L1734.92 726.851L1739.51 723.379Z"
-        fill="var(--color)"
-        stroke="var(--color)"
+        fill="currentColor"
+        stroke="currentColor"
       />
     </svg>
   );
