@@ -1,17 +1,16 @@
 import React from "react";
 
 const Logo = () => {
-    return (<svg width="200" height="64" viewBox="0 0 200 64" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M35.4003 38.5434L37.1089 36.8684C36.4381 37.3131 35.8582 37.8816 35.4003 38.5434V38.5434Z" fill="currentColor"/>
-<path d="M64.0235 63.7845H60.6365L36.6985 39.8549V63.7845H34.268V42.1916C34.2637 40.8914 34.6564 39.6209 35.3937 38.55V38.55L37.1022 36.875C38.1534 36.1785 39.3873 35.809 40.6483 35.813H61.4372V38.2452H38.4791L64.0235 63.7845Z" fill="currentColor"/>
-<path d="M104.298 63.7896H74.8962C74.0566 63.7925 73.2248 63.6292 72.4486 63.3093C71.6724 62.9893 70.9672 62.519 70.3735 61.9253C69.7799 61.3317 69.3095 60.6264 68.9896 59.8502C68.6696 59.074 68.5064 58.2422 68.5093 57.4027V54.9705C68.5062 54.1308 68.6692 53.2988 68.9891 52.5224C69.3089 51.746 69.7793 51.0406 70.3729 50.4468C70.9666 49.8529 71.6719 49.3824 72.4482 49.0624C73.2245 48.7423 74.0565 48.579 74.8962 48.5819H101.864V38.2419H69.8275V35.808H97.9113C98.7508 35.8051 99.5826 35.9684 100.359 36.2883C101.135 36.6083 101.84 37.0786 102.434 37.6723C103.028 38.2659 103.498 38.9712 103.818 39.7474C104.138 40.5236 104.301 41.3554 104.298 42.195V63.7896ZM101.864 61.3558V51.0157H70.9431V61.3558H101.864Z" fill="currentColor"/>
-<path d="M146.878 57.4027C146.881 58.2423 146.717 59.0741 146.397 59.8503C146.077 60.6265 145.607 61.3317 145.013 61.9254C144.42 62.519 143.715 62.9894 142.938 63.3093C142.162 63.6293 141.33 63.7925 140.491 63.7896H112.409V35.8164H140.491C141.33 35.8136 142.162 35.9768 142.938 36.2968C143.715 36.6167 144.42 37.0871 145.013 37.6807C145.607 38.2744 146.077 38.9796 146.397 39.7558C146.717 40.532 146.881 41.3638 146.878 42.2034V57.4027ZM144.446 61.3558V38.2402H114.841V61.3558H144.446Z" fill="currentColor"/>
-<path d="M157.422 63.7896H154.99V35.8164H157.422V63.7896Z" fill="currentColor"/>
-<path d="M200 57.4027C200.003 58.2423 199.839 59.0743 199.519 59.8506C199.199 60.6269 198.729 61.3322 198.135 61.9259C197.541 62.5196 196.836 62.9899 196.059 63.3098C195.283 63.6296 194.451 63.7927 193.611 63.7896H171.92C171.08 63.7925 170.248 63.6292 169.472 63.3093C168.696 62.9893 167.991 62.519 167.397 61.9253C166.803 61.3317 166.333 60.6264 166.013 59.8502C165.693 59.074 165.53 58.2422 165.533 57.4027V42.195C165.53 41.3554 165.693 40.5236 166.013 39.7474C166.333 38.9712 166.803 38.2659 167.397 37.6723C167.991 37.0786 168.696 36.6083 169.472 36.2883C170.248 35.9684 171.08 35.8051 171.92 35.808H193.615C194.454 35.8054 195.286 35.9688 196.062 36.2888C196.838 36.6089 197.543 37.0793 198.136 37.6729C198.73 38.2665 199.2 38.9717 199.52 39.7478C199.84 40.5239 200.003 41.3555 200 42.195V57.4027ZM197.566 61.3558V38.2402H167.966V61.3558H197.566Z" fill="currentColor"/>
-<path d="M70.9983 28.4262H41.6985C40.6997 28.4276 39.7104 28.2318 38.7874 27.8502C37.8643 27.4686 37.0257 26.9086 36.3194 26.2023C35.6131 25.4961 35.0531 24.6574 34.6715 23.7343C34.2899 22.8113 34.0941 21.822 34.0955 20.8232V8.04769C34.0939 7.04873 34.2895 6.05929 34.671 5.13606C35.0526 4.21283 35.6125 3.37396 36.3188 2.66751C37.0251 1.96107 37.8638 1.40093 38.787 1.01921C39.7101 0.637486 40.6995 0.441676 41.6985 0.443H63.3936C64.3925 0.441677 65.3818 0.63742 66.305 1.01902C67.2281 1.40062 68.0669 1.96057 68.7733 2.66682C69.4797 3.37307 70.0398 4.21173 70.4216 5.13477C70.8034 6.05782 70.9994 7.04712 70.9983 8.04601V16.8685H38.9615V23.5586H70.9983V28.4262ZM66.1323 12.0008V5.31067H38.9615V12.0008H66.1323Z" fill="currentColor"/>
-<path d="M113.477 20.8232C113.478 21.822 113.283 22.8111 112.901 23.7341C112.519 24.657 111.959 25.4956 111.253 26.2018C110.547 26.9081 109.708 27.468 108.785 27.8497C107.862 28.2314 106.873 28.4273 105.874 28.4262H76.5745V23.5586H108.611V16.8685H82.8609C81.862 16.87 80.8725 16.6744 79.9493 16.2929C79.0261 15.9114 78.1872 15.3514 77.4807 14.6451C76.7743 13.9388 76.2142 13.1001 75.8324 12.1769C75.4507 11.2538 75.2549 10.2644 75.2562 9.26544V8.04769C75.2549 7.04865 75.4507 6.05918 75.8324 5.13594C76.2141 4.21269 76.7742 3.37383 77.4806 2.66741C78.1871 1.96099 79.0259 1.40088 79.9492 1.01918C80.8724 0.637475 81.8619 0.441677 82.8609 0.443H112.159V5.31067H80.1222V12.0008H105.874C106.873 11.9997 107.862 12.1956 108.785 12.5773C109.708 12.959 110.547 13.519 111.253 14.2252C111.959 14.9314 112.519 15.77 112.901 16.6929C113.283 17.6159 113.478 18.605 113.477 19.6038V20.8232Z" fill="currentColor"/>
-<path d="M32.0302 28.4161H25.2697L4.86769 8.03255V28.4262H2.33798e-05V8.0493C-0.00331583 6.70603 0.351093 5.38611 1.02682 4.22518C1.02971 4.21803 1.03431 4.2117 1.04022 4.20675C1.18588 3.95073 1.34814 3.70453 1.52599 3.46973C2.14517 2.64774 2.92472 1.95993 3.81744 1.44796C4.07641 1.29922 4.34374 1.16555 4.61811 1.04762L4.63989 1.03757L4.75547 0.992348C5.66001 0.625797 6.62704 0.43808 7.60304 0.439584H29.6047V5.3106H8.90789L32.0302 28.4161Z" fill="currentColor"/>
-</svg>)
+  return (<svg width="179" height="56" viewBox="0 0 179 56" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M57.1382 55.9946H54.1084L32.7528 34.8415V55.9946H30.582V36.9061C30.582 35.7035 30.953 34.5975 31.587 33.6868L31.5926 33.6813L33.1173 32.1987C34.02 31.6047 35.1077 31.2594 36.2819 31.2594H54.8298V33.4096H34.3398L57.1373 55.9937L57.1382 55.9946Z" fill="currentColor" />
+    <path d="M93.0833 55.9992H66.8441C63.6777 55.9992 61.1443 53.4899 61.1443 50.3525V48.2014C61.1443 45.064 63.6777 42.5547 66.8441 42.5547H90.9125V33.4133H62.3203V31.2622H87.3834C90.5499 31.2622 93.0833 33.7715 93.0833 36.908V55.9974V55.9992ZM90.9116 53.8481V44.7067H63.3151V53.8481H90.9116Z" fill="currentColor" />
+    <path d="M131.085 50.3533C131.085 53.4898 128.552 56 125.385 56H100.322V31.2639H125.385C128.552 31.2639 131.085 33.7733 131.085 36.9097V50.3533ZM128.914 53.8489V33.415H102.493V53.8489H128.914Z" fill="currentColor" />
+    <path d="M140.495 55.9993H138.324V31.2632H140.495V55.9993Z" fill="currentColor" />
+    <path d="M178.498 50.3533C178.498 53.4898 175.964 56 172.797 56H153.434C150.267 56 147.734 53.4907 147.734 50.3533V36.9097C147.734 33.7733 150.267 31.2639 153.434 31.2639H172.797C175.963 31.2639 178.498 33.7733 178.498 36.9097V50.3533ZM176.326 53.8489V33.415H149.906V53.8489H176.326Z" fill="currentColor" />
+    <path d="M63.3639 24.737H37.2149C33.4599 24.737 30.4292 21.735 30.4292 18.0157V6.72318C30.4292 3.00383 33.4599 0.00183105 37.2149 0.00183105H56.5782C60.3331 0.00183105 63.3639 3.00475 63.3639 6.72318V14.5201H34.7717V20.4357H63.3639V24.7379V24.737ZM59.0204 10.2178V4.30316H34.7717V10.2188H59.0204V10.2178Z" fill="currentColor" />
+    <path d="M101.276 18.0157C101.276 21.735 98.2455 24.737 94.4905 24.737H68.3416V20.4348H96.9337V14.5192H73.9512C70.1963 14.5192 67.1655 11.5172 67.1655 7.79782V6.72226C67.1646 3.00383 70.1963 0.00183105 73.9512 0.00183105H100.1V4.30408H71.508V10.2197H94.4905C98.2455 10.2197 101.276 13.2226 101.276 16.941V18.0166V18.0157Z" fill="currentColor" />
+    <path d="M28.5866 24.7315H22.5521L4.34344 6.70938V24.737H0V6.72411C0 5.49108 0.330964 4.33632 0.91666 3.34364C0.919449 3.33811 0.922238 3.33259 0.927816 3.32706C1.0589 3.10053 1.20207 2.88321 1.36197 2.67602C1.91327 1.95222 2.6096 1.34353 3.40633 0.891391C3.63503 0.76155 3.87489 0.642759 4.12032 0.537782C4.1259 0.535019 4.13427 0.532256 4.14077 0.529494C4.17517 0.515681 4.20957 0.500947 4.24397 0.489897C5.02583 0.172201 5.88856 0 6.7857 0H26.4223V4.30225H7.95058L28.5866 24.7315Z" fill="currentColor" />
+  </svg>)
 }
 
 export default Logo;
